WhatsApp: apply send retry in production plugin outbound path
