+++ extensions/whatsapp/src/outbound-adapter.ts
@@ -1,78 +1,18 @@
-import { resolveAccountEntry } from "openclaw/plugin-sdk/account-resolution";
 import {
   type ChannelOutboundAdapter,
   createAttachedChannelResultAdapter,
   createEmptyChannelResult,
 } from "openclaw/plugin-sdk/channel-send-result";
-import type { OpenClawConfig } from "openclaw/plugin-sdk/config-runtime";
-import {
-  formatErrorMessage,
-  resolveRetryConfig,
-  retryAsync,
-  type RetryConfig,
-} from "openclaw/plugin-sdk/infra-runtime";
 import { resolveOutboundSendDep } from "openclaw/plugin-sdk/outbound-runtime";
 import {
   resolveSendableOutboundReplyParts,
   sendTextMediaPayload,
 } from "openclaw/plugin-sdk/reply-payload";
 import { chunkText } from "openclaw/plugin-sdk/reply-runtime";
-import { createSubsystemLogger, shouldLogVerbose } from "openclaw/plugin-sdk/runtime-env";
-import { resolveDefaultWhatsAppAccountId } from "./accounts.js";
+import { shouldLogVerbose } from "openclaw/plugin-sdk/runtime-env";
+import { resolveWhatsAppRetryConfig, withWhatsAppSendRetry } from "./outbound-retry.js";
 import { resolveWhatsAppOutboundTarget } from "./runtime-api.js";
 import { sendMessageWhatsApp, sendPollWhatsApp } from "./send.js";
-
-const log = createSubsystemLogger("gateway/channels/whatsapp").child("send-retry");
-
-const WHATSAPP_SEND_RETRY_DEFAULTS = {
-  attempts: 3,
-  minDelayMs: 1_000,
-  maxDelayMs: 30_000,
-  jitter: 0.1,
-} satisfies Required<RetryConfig>;
-
-// Only retry clearly transient network errors to avoid duplicate message delivery.
-const WHATSAPP_SEND_RETRY_RE = /timeout|connect|reset|closed|unavailable|temporarily/i;
-
-function shouldRetryWhatsAppSend(err: unknown): boolean {
-  return WHATSAPP_SEND_RETRY_RE.test(formatErrorMessage(err));
-}
-
-function withWhatsAppSendRetry<T>(
-  fn: () => Promise<T>,
-  label: string,
-  configRetry: RetryConfig | undefined,
-): Promise<T> {
-  const resolved = resolveRetryConfig(WHATSAPP_SEND_RETRY_DEFAULTS, configRetry);
-  return retryAsync(fn, {
-    ...resolved,
-    label,
-    shouldRetry: shouldRetryWhatsAppSend,
-    onRetry: (info) => {
-      const maxRetries = Math.max(1, info.maxAttempts - 1);
-      log.warn(
-        `whatsapp send retry ${info.attempt}/${maxRetries} for ${info.label ?? label} in ${info.delayMs}ms: ${formatErrorMessage(info.err)}`,
-      );
-    },
-  });
-}
-
-// Account-level retry takes precedence; falls back to channel-level.
-// Resolves the effective account id (including the default account) so that
-// channels.whatsapp.accounts.<default-id>.retry is honored even when no
-// explicit accountId is passed by the caller.
-function resolveWhatsAppRetryConfig(
-  cfg: OpenClawConfig,
-  accountId?: string | null,
-): RetryConfig | undefined {
-  const root = cfg.channels?.whatsapp;
-  const effectiveAccountId = accountId?.trim() || resolveDefaultWhatsAppAccountId(cfg);
-  if (effectiveAccountId) {
-    const accountCfg = resolveAccountEntry(root?.accounts, effectiveAccountId);
-    if (accountCfg?.retry !== undefined) return accountCfg.retry;
-  }
-  return root?.retry;
-}
 
 function trimLeadingWhitespace(text: string | undefined): string {
   return text?.trimStart() ?? "";
